showbiz/requirements.txt
asgiref==3.2.7
Django==3.0.6
django-admin-interface==0.12.0
psycopg2-binary==2.8.4
django-filebrowser-no-grappelli==3.8.0
django-tinymce4-lite==1.7.5
Pillow==6.2.1
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
django-debug-toolbar
djangorestframework == 3.11.0
django-filter == 2.3.0



gunicorn==20.0.4
whitenoise==5.0.1
Brotli==1.0.7
